fix: 오류 재수정

+++ src/main/java/com/aivle/sellon/rawdb/entity/ClassifiedItemAspect.java
@@ -28,7 +28,7 @@ public class ClassifiedItemAspect {
     private InquiryType aspect;
 
     @Column(name = "sentiment", nullable = false)
-    private Integer sentiment;
+    private Short sentiment;
 
     @Column(name = "mixed_signal")
     private Boolean mixedSignal;
